Add apiID to Configuration
The API interface ID is needed for the "External Koppelprofiel" base URI.
This URI is https://hostname/public/<api_id>/<object_type>

For our API v1 the header Api-Interface-Id is needed to talk to our API. On
environments where there is one API v1 this header is not needed. When
there are multiple v1 API's configured this header becomes mandatory.