!!! FEATURE: Use API ID in object and v1 client
This adds usage of the API ID to the client objects.

Object client: appends the API ID to the base url
V1 client: sends the API ID using the Api-Interface-Id header

This change is breaking as it now requires to explicitly
set the API ID and in case of the Object API the API ID
has to be removed from given API base url.